Move and rename activities
It's usual that activities are postfixed with the activity word for better readability.
As a beginning of a well-structured project the activities moved under the ui layer package.
