+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name="org.sinisterstuf.guesstheanimal.Greeting"
+            android:name="org.sinisterstuf.guesstheanimal.ui.GreetingActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -19,12 +19,12 @@
             </intent-filter>
         </activity>
 
-        <activity android:name="Guess"/>
-        <activity android:name="Win"/>
-        <activity android:name="Lose"/>
-        <activity android:name="LearnName"/>
-        <activity android:name="LearnQuestion"/>
-        <activity android:name="LearnYesNo"/>
+        <activity android:name=".ui.GuessActivity"/>
+        <activity android:name=".ui.WinActivity"/>
+        <activity android:name=".ui.LoseActivity"/>
+        <activity android:name=".ui.LearnNameActivity"/>
+        <activity android:name=".ui.LearnQuestionActivity"/>
+        <activity android:name=".ui.LearnYesNoActivity"/>
         
     </application>
