+++ app/src/main/java/org/sinisterstuf/guesstheanimal/ui/GreetingActivity.java
@@ -1,11 +1,15 @@
-package org.sinisterstuf.guesstheanimal;
+package org.sinisterstuf.guesstheanimal.ui;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 
-public class Greeting extends Activity {
+import org.sinisterstuf.guesstheanimal.Animal;
+import org.sinisterstuf.guesstheanimal.Game;
+import org.sinisterstuf.guesstheanimal.R;
+
+public class GreetingActivity extends Activity {
 	
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -20,9 +24,9 @@ public class Greeting extends Activity {
 		// reverse-linked list traversal:
 		firstAnimal.noAnimal.prevAnimal = firstAnimal;
 		Game.firstAnimal = firstAnimal;
-		Intent intent = new Intent(this, Guess.class);
+		Intent intent = new Intent(this, GuessActivity.class);
 		intent.putExtra(Animal.ANIMAL, firstAnimal);
-		intent.putExtra(Guess.FINAL_GUESS, false);
+		intent.putExtra(GuessActivity.FINAL_GUESS, false);
 		startActivity(intent);
 	}
