+++ app/src/main/java/org/sinisterstuf/guesstheanimal/ui/LearnNameActivity.java
@@ -1,4 +1,4 @@
-package org.sinisterstuf.guesstheanimal;
+package org.sinisterstuf.guesstheanimal.ui;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -6,7 +6,10 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.EditText;
 
-public class LearnName extends Activity {
+import org.sinisterstuf.guesstheanimal.Animal;
+import org.sinisterstuf.guesstheanimal.R;
+
+public class LearnNameActivity extends Activity {
 	
 	private Animal previous;
 	private boolean prevReq;
@@ -23,7 +26,7 @@ public class LearnName extends Activity {
     public void learn(View view) {
     	EditText textInput = (EditText)findViewById(R.id.userInput);
     	String name = textInput.getText().toString();
-    	Intent intent = new Intent(this, LearnQuestion.class);
+    	Intent intent = new Intent(this, LearnQuestionActivity.class);
     	intent.putExtra(Animal.NAME, name);
     	intent.putExtra(Animal.ANIMAL, previous);
     	intent.putExtra(Animal.NEXT_REQ, prevReq);
