+++ app/src/main/res/layout/greeting.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/padding_default"
-    tools:context=".Greeting" >
+    tools:context=".ui.GreetingActivity" >
 
     <TextView
         android:id="@+id/greetingText"
