+++ app/src/main/res/layout/guess.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/padding_default"
-    tools:context=".Guess" >
+    tools:context=".ui.GuessActivity" >
 
     <TextView
         android:id="@+id/guessText"
