+++ app/src/main/res/layout/learn_name.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/padding_default"
-    tools:context=".LearnName" >
+    tools:context=".ui.LearnNameActivity" >
 
     <TextView
         android:id="@+id/learnText"
