+++ app/src/main/res/layout/learn_yesno.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/padding_default"
-    tools:context=".LearnYesNo" >
+    tools:context=".ui.LearnYesNoActivity" >
 
     <TextView
         android:id="@+id/learnText"
