+++ app/src/main/res/layout/lose.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/padding_default"
-    tools:context=".Lose" >
+    tools:context=".ui.LoseActivity" >
 
     <TextView
         android:id="@+id/loseText"
