+++ app/src/main/res/layout/win.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/padding_default"
-    tools:context=".Win" >
+    tools:context=".ui.WinActivity" >
 
     <TextView
         android:id="@+id/winText"
